Form for

+++ src/components/table-row/index.js
@@ -1,10 +1,14 @@
 import React from 'react';
-import Icon from 'utils/icon';
+import Icon from './../../utils/icon';
 
 class TableRow extends React.Component {
 
   shouldComponentUpdate = (nextProps) => {
     if (nextProps.childPropsHaveChanged) {
+      return true;
+    }
+
+    if (nextProps.gutterFields) {
       return true;
     }
 
@@ -15,23 +19,37 @@ class TableRow extends React.Component {
     var row = [],
         rowID = this.props.row_id;
 
-    if (!this.props.placeholder) {
+    if (!this.props.placeholder && !this.props.gutterFields) {
       row.push(
-        <td key={ rowID + 'actions' } className="ui-table-row__td">
-          <button className="ui-table-row__delete" id={ rowID } onClick={this.deleteMethod}>
+        <td key={ rowID + 'actions' } className="ui-table-row__td ui-table-row__td--actions">
+          <button type="button" className="ui-table-row__delete" id={ rowID } onClick={this.deleteMethod}>
             <Icon type="delete" className="ui-table-row__delete-icon" />
           </button>
         </td>
       );
     } else {
-      row.push(<td key={ rowID + 'actions' } className="ui-table-row__td"></td>);
+      var tdClass = "ui-table-row__td ui-table-row__td--actions";
+
+      if (this.props.gutterFields) {
+        tdClass += " ui-table-row__td--gutter";
+      }
+
+      row.push(<td key={ rowID + 'actions' } className={ tdClass }></td>);
     }
 
     for (var key in this.props.fields) {
       var field = this.props.fields[key];
 
-      if (field) {
-        var value = (this.props.data) ? this.props.data.get(field.props.name) : "";
+      if (this.props.gutterFields) {
+        var gutterField = this.props.gutterFields[field.props.name];
+
+        if (gutterField) {
+          row.push(<td key={ key + "gutter" } className="ui-table-row__td ui-table-row__td--gutter">{ gutterField }</td>);
+        } else {
+          row.push(<td key={ key + "gutter" } className="ui-table-row__td ui-table-row__td--gutter"></td>);
+        }
+      } else {
+        var value = (this.props.data) ? this.props.data.get(field.props.name) : null;
         row.push(this.buildCell(field, value));
       }
     }
@@ -47,17 +65,20 @@ class TableRow extends React.Component {
   buildCell = (field, value) => {
     var rowID = this.props.row_id,
         fieldProps = {
-          value: value,
           label: false,
           key: rowID,
-          name: rowID + field.props.name,
+          name: `[${this.props.name}_attributes][${rowID}][${field.props.name}]`,
           row_id: rowID,
           namespace: this.props.name,
           onChange: this.props.updateRowHandler
         };
 
+    if (value) {
+      fieldProps.value = value
+    }
+
     if (this.props.placeholder) {
-      fieldProps._placeholder = true
+      fieldProps._placeholder = true;
     }
 
     var fieldHTML = React.cloneElement(field, fieldProps);
@@ -71,8 +92,14 @@ class TableRow extends React.Component {
    * @method render
    */
   render() {
+    var mainClasses = "ui-table-row";
+
+    if (this.props.gutterFields) {
+      mainClasses += " ui-table-row--gutter";
+    }
+
     return (
-      <tr className="ui-table-row">
+      <tr className={ mainClasses }>
         { this.buildRow() }
       </tr>
     );
